Generalise the 'get_tinker' method to 'get_param'.
The `get_param` method fetches a value from the configuration by keyword and
optionally by parameter, with an optional default in case the keyword or
parameter are not found.  Keywords such as 'rtcsync' can be fetched directly,
as can keywords without parameters like 'includefile' and keywords with
parameters such as 'tinker'.

Full tests and documentation improvements included.