Test the documentation to make sure the example code works from the input data
Insights parsers and combiners usually contain samples of the input data and
code examples that are written as if a user had imported the parser or combiner
and was trying commands out in an interpretive shell.  Not only is it useful
to make sure that this code actually works and demonstrates the correct usage
of the class under test, but the code also forms useful test code that should
be checked by the test system.

We introduce the `test_doc_examples` function, which tests all the classes in a
given module.  If the parser module given only contains one parser, the
documentation for it is assumed to be in the parser module.  If there is more
than one parser, the module documentation is ignored and the documentation of
the parser itself is assumed to contain the input data and examples.

This work is done by the `test_examples_in_doc_of` function, which looks in the
documentation given for an indented block (following a ``\:\:``) and an
`Examples\:` section.  The indented block is stripped of leading space and read
as test data, and put in a 'shared' dictionary against the given parser using
the `'context_wrap'` function.  This is stored in a dictionary of local
declarations, to be used in later evaluation.

Each example line preceded by '`>>> `' is assumed to be a command and
`eval`uated, and the result is asserted to be equal to any non-'>>> ' indented
lines (also `eval`'ed).  The `'variable' = shared['parser']` declaration is
used to set the given name in the local declarations dictionary.

Some attempt is made to set up the local evaluation environment but it is not
assumed to be a complete Python environment.  In particular, other statements
that are not 'variable = value' or 'expression' are not parsed.  Only 'variable
= value' statements update the local environment.