Change to providing get_param and get_last methods
From the suggestion by @xiangce, the `get_param` method now lists all of the
available definitions of a given keyword and parameter, and the new `get_last`
method selects the last of those.  Defaults and parameter lookup are as before.
This allows rule writers to optimistically select the last, or check for all
definitions if being more cautious.

Documentation and tests are also provided.